Fixed some bugs in the lighting system and added specular light
the step variable which tells us our position between two intersections
was int, thus causing data loss.
Fixed some of the vector directions as well. This kinda works now

+++ CGWork/IritLight.h
@@ -26,4 +26,4 @@ struct light {
  * @state				- contains all light sources in our world
 */
 RGBQUAD calculateLight(struct IntersectionPoint &intersecting_x1, struct IntersectionPoint &intersecting_x2,
-					   int t, struct State &state);
+					   double t, struct State &state);
